refactor: nannka yoku natta
